(FEAT) Add custom_powershell_property util function
This commit adds a new utility method, custom_powershell_property.
This method takes a name and an expression and returns a string which
represents a PowerShell custom object property hash.

The expression being passed must already be an interpolated string as
this method does not do any additional work for interpolation; whatever
is passed is what will end up as PowerShell code in the expression
script block.