(MAINT) Conform hash util methods
This commit standardizes the hash util methods on taking object
as a parameter as they handle more data types than just hashes.

It also conforms the hash util methods to use the same underlying
structure by utilizing a new private method, apply_key_mutator,
to do the heavy lifting. The other hash utility methods now create
a proc for the conforming of the keys and pass that to the mutator.