(MAINT) Handle symbols with case methods
This commit updates the casing methods to handle symbols
if passed them - they will return a string if given a
string and a symbol if given a symbol, properly cased.

This does away with the weird requirement to cast the
keys to strings, case them, and resymbolize them in the
hash mutation procs.