(MODULES-10389) Only load pwshlib if functionality present
This is a hotfix to resolve an issue customers are seeing after the v.3.0.0 release of the puppetlabs-powershell module. Adding a pwshlib feature will ensure that the powershell provider is only loaded if the functionality is detected. On Linux systems, this should prevent Puppet from throwing an error when it cannot load ruby-pwsh