(IAC-887) avoid test failures on new powershell 7 behaviour
This likely is a problem with the tests not reacting well to a
behavioral change in powershell 7, and not a code issue. The same
tests ran fine previously with 6.2.4-1.ubuntu.18.04 in
https://travis-ci.com/github/puppetlabs/ruby-pwsh/jobs/293023278
while the same commit (https://github.com/puppetlabs/ruby-pwsh/commit/03e90c7d582c7391cca09bd6304d2be14764ad18) fails on powershell 7.

This behaviour is also linux-specific.