(IAC-1045) Move DSC base provider to pwshlib
Prior to this commit the code for the base provider of the new
puppetized DSC modules was vendored into each module; this
meant that if the provider ever needs to be updated, *every*
prior release of the puppetized module would need to be built
again and pushed with an incremented number.

Moving the base provider to the pwshlib Puppet module (which
is this ruby library packaged as a Puppet module) means that
the provider can be updated separately from the PowerShell
builder code and does not require rebuilding puppetized modules,
only updating the reference pin for pwshlib.

This also makes testing the provider easier as it lives in a
ruby library now instead of in a PowerShell module.