(GH-55) Handle intentionally empty arrays
Prior to this commit the DSC base provider would
occasionally need to send an empty array as a
defined property to Invoke-DscResource. However,
the CIM instance on the other end may not be able
to appropriately cast from an empty array to, say
a [String[]] type; this would cause the run to
fail. Furthermore, the provider did not have a way
to distinguish between an *absent* setting and an
*empty* one. Ruby does not treat empty arrays and
nil as the same, even though PowerShell can/does.

This commit adds a special handler for munging the
data returned from a get invocation, ensuring that
if a propertys value is nil AND the property was
used for the query AND the resource expects an
array, THEN it will set the result to the manifest
value or an empty array.

It also strongly types any empty arrays being sent
to Invoke-DscResource, as this prevents the casting
problem by making the array type casting  explicit
instead of merely implicit.

- Resolves #55