(GH-59) Refactor away from Simple Provider
Prior to this commit the DSC Base provider inherited from
the Resource APIs Simple Provider. This brought with it a
limitation in the type definition, requiring all of the
generated DSC types to have the ensure attribute. This
led to the confusion in puppetlabs/Puppet.Dsc#72 as well
as necessitated the filing of puppetlabs/Puppet.Dsc#74
to improve the UX.

This commit removes the dependence on the Simple Provider
and instead copies over the code from it - namely, the
set method - and modifies it slightly for brevity.

This change sets up a few things:

1. The collapse of the ensure and dsc_ensure attributes.
   Once merged, the ensure attribute becomes redundant;
   the provider no longer needs to care about it and it
   can be removed from both generation and the provider.
2. The dropping of the simple_get_filter from puppetized
   DSC Resource type definitions. The feature is made
   redundant in this commit and no longer used.
3. Refactoring of this provider for clarity. Now that all
   of the required code for this provider lives in this
   gem we can look to refactor some of the helpers out and
   clean things up a little.

- Resolves #59
- Unblocks puppetlabs/PUppet.Dsc#74