(MAINT) Fix ensurability in dsc base provider
Prior to this commit the dsc base provider erroneously only
tried to compare ensurable states on resources which are NOT
ensurable instead of only comparing them on resources which
ARE ensurable.

This commit corrects that error and adds special handling
in invoke_get_method for trimming empty properties when
returning the representation of an absent resource on the
system in order to prevent flapping during Puppet runs.