(MAINT) Instantiate variables after first use
Prior to this commit the variables for credentials and cim instances
were instantiated prior to the definition of the block for that
credential or instance; this caused the variable interpolation step
to replace the value definition with a self-referencing variable which
would always be null, and thus fail to correctly instantiate.

This commit moves the code for variable instantiation to *after* the
credential or CIM instance is created, meaning only subsequent calls
will reference the variable.