(MAINT) Correctly canonicalize enumerable values in dsc
Prior to this commit the comparison in the DSC base provider for
values on the node vs values in the manifest was a simple equals
check; this works for non-enumerables and for enumerables *so long
as* the enumerables are in the same order.

This commit corrects the check to sort the enumerables so that the
comparison is done purely on the values, not the order, as PowerShell
does not care about the order of items in a hash or array.