(GH-105) Ensure canonicalized_cache check validates against namevar
Prior to this commit the safety check in the canonicalize method intended to prevent
multiple executions to canonicalize the same resource compared the cached resources
to the *full* resource being checked; this commit changes that check to only search
for a hash containing the matching namevars.

This should prevent multiple runs and bugs where mandatory parameters are dropped
during an RSAPI refresh event.