(GH-105) Ensure set runs on ambiguous ensure states
Prior to this commit if a resource is ensurable but the value of ensure on
the target node was not Present or Absent the call to run Invoke-DscResource
never gets made.

This commit updates the logic in the set method to call update() if the state
of the ensurable resource is ambiguous from a CRUD perspective, which is
preferable to not running set at all.