(MAINT) Add fallback cache check in get
This commit adds a fallback check for the canonicalized cache in the get
method of the dsc base provider; malformed DSC Resources which are not
ensurable may be returned such that the namevar/required properties that
are discovered on disk do not match the ones in the manifest; this happens
because DSC does not actually always treat properties marked as required in
precisely the same way Puppet treats namevars.

This commit therefore checks the canonicalized hash to see if a result for
that resource was found and has a matching *Puppet* name, ignoring all other
keys in the hash for the comparison.