(GH-145) Add insync? and invoke_test_method to dsc provider
Prior to this commit the DSC base provider had no way to handle resource
level insync checks and could only rely on the default property-by-property
insync checks in Puppet::Property. This fell short in cases where such a
check could not work, such as in cases where the DSC Resource does not
return the expected values in Get or had a mismatch between its API spec
and the implementation.

This commit:

1. Adds a class variable, `@@cached_test_results`
2. Adds an accessor method, `cached_test_results`, to retrieve the cache
3. Adds the `insync?` method which can be called by the resource API;
   for more info, see puppetlabs/puppet-resource_api#225. This method
   only handles custom insync validation for a resource which has the
   `validation_mode` specified as `resource` and otherwise defaults to
   the existing behavior. If the `validation_mode` *is* specified as
   `resource`, it instead first checks the test result cache for
   whether or not the resource is in the desired state and, if the
   resource has not already been cached, calls `invoke_test_method` and
   returns that result.
4. Adds the `invoke_test_method` method which passes the `should`
   hash of the resource to `Invoke-DscResource` similarly to the extant
   `invoke_set_method`; unlike with `set`, however, it caches the name
   and whether or not the resource is in the desired state and then it
   returns `true` if in the desired state and otherwise `false` with a
   descriptive change message. In cases where something goes wrong, it
   returns `nil` and causes Puppet to fall back on the default property-
   by-property comparison logic.

This commit includes new unit tests for the additional methods.

This commit makes progress towards but does not alone implement
puppetlabs/Puppet.Dsc#145.