(MODULES-11051) Refactor PSHost Env Var Reset
This commit refactors the Reset-ProcessEnvironmentVariables function
and calling in the PowerShell host init script to cache the
environment variables when the host is started up by just enumerating
them and then using that cached array to reset them inside the host.

This ensures that the value of the environment variable matches those
outside of the PowerShell Host instead of using expected inheritance;
this resolves the issue where a process scope environment variable does
not match the expected User > Machine inheritance order, as with the
TEMP variable under the System user.

This also simplifies the overall code. All existing unit tests continue
to pass with this change.