(GH-127) Canonicalize enums correctly
Prior to this commit, some enums would be incorrectly canonicalized to the
value returned from invoke_get_method; while this is fine for most strings,
enums are case sensitive and validated in Puppet. For example, when the
result of invoke_get_method returns a value of DWord for an enum of Dword,
an error was raised when attempting to canonicalize to the incorrect casing
of DWord.

This commit:

1. Adds a new enum_attributes method which, like the namevar_attributes and
   parameter attributes methods, filters the type attributes to return the
   list of attribute names (as keys); in this case, for any attribute whose
   data type includes an Enum.
2. Ensures that enums are not erroneously cased to match system state.
3. Includes new unit tests for the canonicalize behavior as well as the new
   enum_attributes method.