(GH-137) Add explanatory comments to init.ps1
Paired on this with James Pogran to figure out what most of the
invocation-specific code is doing - this commit does not take
care to thoroughly document the commands and behavior around
the Pipe Server or the C# code but does annotate the behavior
most likely to be modified for functionality.

When this script is broken up into component pieces, the
individual functions should get full comment-based help
and the extracted C# code should get full reference docs.