(GH-136) Ensure init.ps1 C# classes namespace to RubyPwsh
Prior to this commit, the C# code which is used for the custom
PowerShell host referenced Puppet; this is a holdover from when
the code was included in the PuppetLabs PowerShell module and
was explicitly Puppet-only. Now that the code is vendor-agnostic,
the references have been cleaned up to refer to the parent
project, ruby-pwsh, instead.

Resolves #136