(GH-134) Move C# code to separate file
Prior to this commit, the CSharp code for the custom PowerShell host
used in the PowerShell manager lived as a here string in the init
script; this was clunky, difficult to maintain, and untestable.

This commit moves that code to a new C# file, which enables easier
reading, shortens the init script considerably, and paves the way
for future unit testing.