(GH-143) Handle order insensitive arrays in same?
Prior to this commit, a typo in the case statement of the same? method
in the DSC Base Provider caused the Hash and Array cases never to run,
always falling back on a simple == value comparison.

This caused arrays which were the same except for ordering to always
fail, as == comparisons of arrays are order sensitive.

This commit corrects the typo, ensuring that the Array case runs. This
only handles order insensitivity of arrays which have no nested values.

A future commit will be required to deeply sort arrays and hashes - see
#144 for further details.

This commit also comments out the Hash case, which otherwise errors. This
preserves existing behavior for Hashes, causing them to be compared with
==, which is order insensitive for _keys_ but not for array values (this
also will be resolved when #144 is implemented.

- Resolves #143