(GH-144) Enable order-insensitive comparisons for DSC
Prior to this commit value comparisons in the DSC Base Provider
only partially worked for arrays and hashes; in PowerShell, the
order of a hash or array does not matter for comparison but in
Ruby hashes and arrays are order-sensitive.

The same? method was meant to get around this but only worked
for non-enumerable values and values which were simple arrays.

This commit adds a new recursively_sort method to obviate this
issue and calls it in same?