(MAINT) Refactor secrets handling in DSC
Prior to this commit, the new handling for secrets was not utilized when
preparing credentials for PowerShell. Further, the initial implementation
introduced a severe performance hit when attempting to use gsub against
a 150+ line block of text.

This commit ensures that credential secrets are handled like all other
secrets. It also refactors the redact_secrets & remove_secret_identifiers
methods to use a helper method, handle_secrets.

This new handle_secrets method takes an input string, replacement string,
and error message as parameters. It parses the input string, first looking
to see if any secrets are in it. If not, it returns the input string. Next,
it breaks the input string into separate lines - the original gsub regex
was causing noticeable performance impacts when run against the full text.

Instead, it now first naively checks to see if a line includes a secret and,
if not, returns the line. If it *does* include a secret, that secret is then
substituted based on the supplied replacement string from the method params.

The modified text is then rejoined with newlines and, if any secret ids remain,
an error is raised instead of leaking secrets or misapplying PowerShell.

Finally, the method returns the modified text.