(GH-154) Fix canonicalization in get for DSC
Prior to this commit the call to canonicalize for a retrieved resource
would return the previously cached canonical value for the specified
name hash; in other words, it would return the canonicalized *should*
value. The strict check for retrieved data from get validates that
calling canonicalize does *not* change the result, as it is meant to
munge the manifest input into a state that matches what get will return.

We have been (ab?)using this functionality to force the manifest values
to be case insensitive but case preserving. We do *not* want to munge
the return data from the get method.

This commit modifies the logic for handling canonicalizing a resource
hash to only munge if no entry for the namehash was found in the cache;
otherwise, it returns the data as-is.

This can *only* happen when a resource has already had its should hash
canonicalized during catalogue compilation and a strict check is being
applied to the data returned from get.