(GH-154) Fix DSC returns for empty strings
Prior to this commit, the ConvertTo-CanonicalResult function used
to turn output from Invoke-DscResource with the Get method into
something Puppet can use would return empty strings for unset
values.

Puppet distinguishes between nil and an empty string when deciding
if a value has been defined - in these cases, nil is treated as
undef whereas an empty string is treated as a set value.

This causes canonicalization (and the run) to fail when the value
for an unset enum is returned as an empty string.

This commit modifies the logic in ConvertTo-CanonicalResult to
munge an empty string to nil to prevent invalid enum errors.