(GH-154) Fix DSC returns for CIM instance arrays
Prior to this commit, the logic for determining if a CIM Instance
value should be munged to an array value failed against hashes;
the logic checked to see if the count was less than 2 and, if so,
wrapped the value in an array.

This failed against hashes because the count property represents
the number of key-value pairs in a hash; this means that to the
original logic, a single hash with multiple keys would never be
wrapped in an array.

When passing back a single item for a value to Puppet which Puppet
expects to be an array, an error gets raised.

This commit corrects the logic by validating the typename of the
value includes the [] suffix and thereby ensures that values
returned for CIM instances which are single hashes and should be
treated as single-member arrays of hashes can be canonicalized.