(GH-67) Move CI to Github Actions
Prior to this commit, the CI for this project was split across travis
and appveyor to ensure coverage for Windows and Linux nodes alike.

This commit replaces both travis and appveyor with a single GHA workflow
with two jobs: rubocop and spec. This ensures that the project is tested
against our supported suite without requiring multiple CI config files.