(GH-163) Ensure acceptance tests run in CI
This commit adds a job for running acceptance tests to the GitHub Actions
CI workflow. It modifies the Gemfile to use a specified Puppet version
if one is set as the environment variable `PUPPET_GEM_VERSION`. If that
environment variable is not set, it defaults to latest (the previous
behavior).

For GitHub Actions, the matrix runs against all supported platforms and
Puppet versions (with the appropriate Ruby version). It was necessary
to also turn on the HTTP listener for WinRM so DSC could work.