(MAINT) Update ci workflows
This commit fixes a few bugs and also refreshes the workflows so they
are inline with the reusable workflow templates.

Note: this repo is an exception and cannot use the reusable workflows
because it contains a library and a Puppet module.

To avoid duplication and conflicts the workflows are defined explicitly
here.