(MAINT) Gemspec explicit file declarations
Prior to this commit spec.files used git to collect the required files
for the gem. This would pull in everything that had been checked in to
git.

We can actually be a little bit more explicit here. This commit updates
spec.files to contain a known list of files to include in the gem.