(CAT-1617) - Always load vendored module in PSModulePath
Prior to this commit, the vendored dsc module would only be loading into
the PSModulePath Env variable if the dsc resource type was implemented
as a Class.

Now that DSC Modules seem to be taking the approach of abstracting the
DscResource.Base module into a seperate module which is then imported,
we need to alter the PSModulePath Env variable to include the vendored
modules path to ensure these modules are imported as expected.