(CAT-1724) - Fix Provider returned data not matching Type Schema
Prior to this commit, we would see the above error returned when dsc
would return a nil value for a required string in its data hash.

As puppet is not expecting a nil value, but rather an empty string, the
error is raised on each puppet run where this occurs.

We can fix this by converting all required string attributes values in
the data hash from nil to an empty string ('').