(CAT-1869) - Add configurable dsc_timeout
This commit introduces a new instance variable @timeout, set
to nil by default. A new method set_dsc_timeout is added to
handle the setting of @timeout based on the name_hash[:dsc_timeout]
value. This timeout specifies the number of seconds the ps_manager will
await the dsc invocation to return, before moving on.

The ps_manager.execute method call now includes the timeout value.
The error message for when output is nil has been updated to include
information about the timeout, if it is set.

Does not affect previous behaviour, therefore is completely backwards
compatible with previous versions.