(CAT-1991) - Skip missing dirs invalid_dir method
Prior to this commit, the invalid_directories? method in pwsh::util
would return true if a directory did not exist. We found this to cause
issues with updating packages which modifiy the LIB env, as missing
entries during install/uninstall/upgrade would raise an error and exit.

Now, we alter the check to skip entries that do not exist, and only
flag for an invalid directory if the path supplied is not a directory.